third_party/WebKit/LayoutTests/http/tests/devtools/profiler/heap-profiler-profiling-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that sampling heap profiling works.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testProfiling
found pageFunction heap-profiler-profiling.js:14
found (anonymous) heap-profiler-profiling.js:15

Profiler was disabled.
